Refactor: extract `ErrorResponse` into its own file
